Seems fundamental protocol parameters are settled
Refuse having sorted/delta option encoding from CoAP (options are now
fully independent from each and can go in any order, as in SCTP, just
lower number/length could be encoded with fewer bytes) and set maximum
number for user streams to 1024 (10 bits). We are protocol for IoT and
big Internet, not for Web frontend/backend cluster, so for client
application 1024 streams is a very serious value even nowadays, decide
that's this is enough for version 0 of protocol. Control data for one
message is also limited to one byte, alas.

There is still a lot of designing work, but order of magnitude for
fundamental sizes is now known. Decouple stream reparent opcode from
window setting, however, to be able to have more bits for stream in far
future.