New thoughts on CBAR (after Tcl and CBOR-packed)
An already given example with Perl's Compress::LZF shows INTs (2 byte)
instead of strings, what if make two passes and feed LZF with such
a preprocessed CBOR? Then:
* Compress::LZF code could be used as is
* and it's format has 8K backreferences instead of 4K (less literals,
  though)
* and number of atoms in preprocesses format coulde be more than 4096

If combined with some of draft-ietf-cbor-packed ideas, it may be
compressed even to single byte for some atoms and perhaps more
appropriately be called something other than CBAR. However, as this
likely could give atoms more typing than just raw bytestring (e.g.
at least flag for prefixing with CBOR length or not, if not CBOR types
itself), then CBAR idea as a "more compact less-typed Tcl-list-like",
based on nibbles and SQLite record format, may still be useful for
InternAtom - it already has a need for a create-if-not-exist flag, and
desire to save 1 byte in "one atom only" vs "many atoms at once"
version, conflicting with desire to use CBOR for it also (but that
would be too fat), and customary bit-in-byte format looks not beatiful
enough.